feat: :construction: new project controller
