+++ apps/server/src/controllers/project.js
@@ -1,30 +1,93 @@
-import { getLogInfos } from '../utils/logger.js'
 import {
-  addRepo,
-  addUser,
-  getUserProjectById,
-  removeUser,
-  updateProjectStatus,
-} from '../models/project-queries.js'
-import {
-  projectFailed,
-  projectCreated,
   getUserProjects,
   projectInitializing,
-} from '../models/project-queries2.js'
+  projectCreated,
+  projectFailed,
+  getProjectById,
+  projectLocked,
+  projectUnlocked,
+  projectAddUser,
+  projectRemoveUser,
+  projectArchiving,
+} from '../models/project-queries.js'
+import { getLogInfos } from '../utils/logger.js'
 import { send200, send201, send500 } from '../utils/response.js'
 import { ansibleHost, ansiblePort } from '../utils/env.js'
-import { getProjectRepositories } from '../models/repository-queries.js'
+import { getUserByEmail, getUserById } from '../models/users-queries.js'
+
+// GET
+
+export const getUserProjectsController = async (req, res) => {
+  const userId = req.session?.user?.id
+
+  try {
+    const projects = await getUserProjects(userId)
+    // TODO : pourquoi ne pas juste renvoyer projects ?
+    // const projectsOnly = await getUserProjects(userId)
+    // const projectsWithRepos = await Promise.all(projectsOnly.map(async (project) => {
+    //   const repos = await getProjectRepositories(project.id)
+    //   return {
+    //     ...project,
+    //     projectName: project.name,
+    //     orgName: project.organization,
+    //     usersId: project.ownerId === userId ? project.usersId : [userId],
+    //     repos,
+    //   }
+    // }))
+    req.log.info({
+      ...getLogInfos(),
+      description: 'Projects successfully retreived',
+    })
+    await send200(res, projects)
+  } catch (error) {
+    const message = 'Cannot retrieve projects'
+    req.log.error({
+      ...getLogInfos(),
+      description: message,
+      error: error.message,
+    })
+    send500(res, message)
+  }
+}
+
+export const getProjectByIdController = async (req, res) => {
+  const projectId = req.params?.id
+  const userId = req.session?.user?.id
+
+  try {
+    const project = await getProjectById(projectId)
+    if (!project.usersId.includes(userId) && project.ownerId !== userId) throw new Error('Requestor is not member of the project')
+
+    req.log.info({
+      ...getLogInfos({ projectId }),
+      description: 'Project successfully retrived',
+    })
+    send200(res, project)
+  } catch (error) {
+    const message = 'Cannot retrieve project'
+    req.log.error({
+      ...getLogInfos({ projectId }),
+      description: message,
+      error: error.message,
+    })
+    send500(res, message)
+  }
+}
+
+// POST
 
 export const createProjectController = async (req, res) => {
   const data = req.body
-  data.status = 'initializing'
   data.ownerId = req.session.user.id
+
   let project
+
   try {
     project = await projectInitializing(data)
     req.log.info({
-      ...getLogInfos({ projectId: project.id }),
+      ...getLogInfos({
+        projectId: project.id,
+      }),
       description: 'Project successfully created in database',
     })
     send201(res, project)
@@ -38,11 +101,11 @@ export const createProjectController = async (req, res) => {
   }
 
   try {
+    const owner = getUserById(project.ownerId)
     const ansibleData = {
       orgName: project.organization,
-      ownerEmail: project.owner.email, // TODO aller chercher la bonne valeur
+      ownerEmail: owner.email,
       projectName: project.name,
-      // envList: project.envList, //virer le requested sur ansible
     }
     await fetch(`http://${ansibleHost}:${ansiblePort}/api/v1/projects`, {
       method: 'POST',
@@ -75,7 +138,7 @@ export const createProjectController = async (req, res) => {
       error,
     })
     try {
-      project = await projectFailed(project)
+      project = await projectFailed(project.id)
 
       req.log.info({
         ...getLogInfos({ projectId: project.id }),
@@ -91,135 +154,28 @@ export const createProjectController = async (req, res) => {
   }
 }
 
-export const addRepoController = async (req, res) => {
-  const userId = req.session?.user?.id
+// PUT
+
+export const projectAddUserController = async (req, res) => {
+  const userId = req.session?.user.id
   const projectId = req.params?.id
   const data = req.body
-  data.status = 'initializing'
 
-  let dbProject
+  let project
   try {
-    dbProject = await getUserProjectById(projectId, userId)
-    if (!dbProject) {
-      const message = 'Missing permissions on this project'
-      req.log.error({
-        ...getLogInfos(),
-        description: message,
-      })
-      return send500(res, message)
-    }
+    project = await getProjectById(projectId)
+    if (!project) throw new Error('Project not found')
 
-    dbProject.locked = true
-    await addRepo(dbProject, data)
-    dbProject = await getUserProjectById(projectId, userId)
+    if (!project.usersId.includes(userId) && project.ownerId !== userId) throw new Error('Requestor is not member of the project')
 
-    const message = 'Git repository successfully added into project'
-    req.log.info({
-      ...getLogInfos({ projectId: dbProject.id }),
-      description: message,
-    })
-  } catch (error) {
-    const message = `Cannot add git repository into project: ${error.message}`
-    req.log.error({
-      ...getLogInfos(),
-      description: message,
-      error,
-    })
-    return send500(res, message)
-  }
-
-  try {
-    const ansibleData = {
-      orgName: dbProject.orgName,
-      ownerEmail: dbProject.owner.email,
-      projectName: dbProject.projectName,
-      internalRepoName: data.internalRepoName,
-      externalRepoUrl: data.externalRepoUrl.startsWith('http') ? data.externalRepoUrl.split('://')[1] : data.externalRepoUrl,
-      isInfra: data.isInfra,
-    }
-    if (data.isPrivate) {
-      ansibleData.externalUserName = data.externalUserName
-      ansibleData.externalToken = data.externalToken
-    }
-
-    await fetch(`http://${ansibleHost}:${ansiblePort}/api/v1/repos`, {
-      method: 'POST',
-      body: JSON.stringify(ansibleData),
-      headers: {
-        'Content-Type': 'application/json',
-        authorization: req.headers.authorization,
-      },
-    })
-
-    try {
-      const indexRepo = dbProject.repos.findIndex(repo => repo.internalRepoName === data.internalRepoName)
-      dbProject.repos[indexRepo].status = 'created'
-
-      dbProject = await updateProjectStatus(dbProject, 'created')
-
-      req.log.info({
-        ...getLogInfos({ projectId: dbProject.id }),
-        description: 'Project status successfully updated in database',
-      })
-    } catch (error) {
-      req.log.error({
-        ...getLogInfos(),
-        description: 'Cannot update repo status to created',
-        error: error.message,
-      })
-      return send500(res, error.message)
-    }
-
-    send201(res, 'Git repository successfully added into project')
-  } catch (error) {
-    const message = 'Provisioning repo with ansible failed'
-    req.log.error({
-      ...getLogInfos(),
-      description: message,
-      error: error.message,
-    })
-
-    try {
-      const indexRepo = dbProject.repos.findIndex(repo => repo.internalRepoName === data.internalRepoName)
-      dbProject.repos[indexRepo].status = 'failed'
-      dbProject = await updateProjectStatus(dbProject, 'failed')
-
-      req.log.info({
-        ...getLogInfos({ projectId: dbProject.id }),
-        description: 'Repo status successfully updated in database to failed',
-      })
-    } catch (error) {
-      req.log.error({
-        ...getLogInfos(),
-        description: 'Cannot update repo status to failed',
-        error: error.message,
-      })
-      return send500(res, error.message)
-    }
-    send500(res, message)
-  }
-}
-
-export const addUserController = async (req, res) => {
-  const userId = req.session?.user?.id
-  const projectId = req.params?.id
-  const data = req.body
-  data.status = 'initializing'
-
-  let dbProject
-  try {
-    dbProject = await getUserProjectById(projectId, userId)
-
-    if (!dbProject) {
-      throw new Error('Missing permissions on this project')
-    }
-
-    await addUser(dbProject, data)
-    dbProject = await getUserProjectById(projectId, userId)
+    const userToAdd = await getUserByEmail(data.email)
+    // TODO : user status initializing ?
+    await projectLocked(projectId)
+    await projectAddUser({ projectId, userId: userToAdd.id })
 
     const message = 'User successfully added into project'
     req.log.info({
-      ...getLogInfos({ projectId: dbProject.id }),
+      ...getLogInfos({ projectId }),
       description: message,
     })
     send201(res, message)
@@ -235,12 +191,11 @@ export const addUserController = async (req, res) => {
   try {
     // TODO : US #132 appel ansible
     try {
-      const indexUser = dbProject.users.findIndex(user => user.email === data.email)
-      dbProject.users[indexUser].status = 'created'
-      dbProject = await updateProjectStatus(dbProject, 'created')
+      // TODO : user status created ?
+      await projectUnlocked(projectId)
 
       req.log.info({
-        ...getLogInfos({ projectId: dbProject.id }),
+        ...getLogInfos({ projectId }),
         description: 'Project status successfully updated in database',
       })
     } catch (error) {
@@ -258,12 +213,11 @@ export const addUserController = async (req, res) => {
       error,
     })
     try {
-      const indexUser = dbProject.users.findIndex(user => user.email === data.email)
-      dbProject.users[indexUser].status = 'failed'
-      dbProject = await updateProjectStatus(dbProject, 'failed')
+      // TODO : user status failed ?
+      await projectUnlocked(projectId)
 
       req.log.info({
-        ...getLogInfos({ projectId: dbProject.id }),
+        ...getLogInfos({ projectId }),
         description: 'Project status successfully updated in database',
       })
     } catch (error) {
@@ -278,42 +232,48 @@ export const addUserController = async (req, res) => {
   }
 }
 
-export const removeUserController = async (req, res) => {
-  const userId = req.session?.user?.id
+// DELETE
+
+export const projectRemoveUserController = async (req, res) => {
+  const userId = req.session?.user.id
   const projectId = req.params?.id
   const data = req.body
 
-  let dbProject
+  let project
   try {
-    dbProject = await getUserProjectById(projectId, userId)
-    if (!dbProject) {
-      throw new Error('Missing permissions on this project')
-    }
+    project = await getProjectById(projectId)
+    if (!project) throw new Error('Project not found')
 
-    await removeUser(dbProject, data)
+    if (!project.usersId.includes(userId) && project.ownerId !== userId) throw new Error('Requestor is not member of the project')
+
+    const userToRemove = await getUserByEmail(data.email)
+    // TODO : user status initializing ?
+    await projectLocked(projectId)
+    await projectRemoveUser({ projectId, userId: userToRemove.id })
 
     const message = 'User successfully removed from project'
     req.log.info({
-      ...getLogInfos({ projectId: dbProject.id }),
+      ...getLogInfos({ projectId }),
       description: message,
     })
-    send200(res, message)
+    send201(res, message)
   } catch (error) {
     const message = 'Cannot remove user from project'
     req.log.error({
       ...getLogInfos(),
-      description: message,
       error: error.message,
     })
     return send500(res, message)
   }
+
   try {
     // TODO : US #132 appel ansible
     try {
-      dbProject = await updateProjectStatus(dbProject, 'created')
+      // TODO : user status deleting ?
+      await projectUnlocked(projectId)
 
       req.log.info({
-        ...getLogInfos({ projectId: dbProject.id }),
+        ...getLogInfos({ projectId }),
         description: 'Project status successfully updated in database',
       })
     } catch (error) {
@@ -331,10 +291,11 @@ export const removeUserController = async (req, res) => {
       error,
     })
     try {
-      dbProject = await updateProjectStatus(dbProject, 'failed')
+      // TODO : user status failed ?
+      await projectUnlocked(projectId)
 
       req.log.info({
-        ...getLogInfos({ projectId: dbProject.id }),
+        ...getLogInfos({ projectId }),
         description: 'Project status successfully updated in database',
       })
     } catch (error) {
@@ -349,56 +310,71 @@ export const removeUserController = async (req, res) => {
   }
 }
 
-export const getUserProjectsController = async (req, res) => {
-  const userId = req.session?.user?.id
+export const projectArchivingController = async (req, res) => {
+  const userId = req.session?.user.id
+  const projectId = req.params?.id
 
   try {
-    const projectsOnly = await getUserProjects(userId)
-    const projectsWithRepos = await Promise.all(projectsOnly.map(async (project) => {
-      const repos = await getProjectRepositories(project.id)
-      return {
-        ...project,
-        projectName: project.name,
-        orgName: project.organization,
-        usersId: project.ownerId === userId ? project.usersId : [userId],
-        repos,
-      }
-    }))
+    const project = await getProjectById(projectId)
+    if (project.ownerId !== userId) throw new Error('Requestor is not owner of the project')
+
+    await projectLocked(projectId)
+    await projectArchiving(projectId)
     req.log.info({
-      ...getLogInfos(),
-      description: 'Projects successfully retreived',
+      ...getLogInfos({
+        projectId,
+      }),
+      description: 'Project successfully archived in database',
     })
-    await send200(res, projectsWithRepos)
+    send201(res, projectId)
   } catch (error) {
-    const message = 'Cannot retrieve projects'
     req.log.error({
       ...getLogInfos(),
-      description: message,
+      description: 'Cannot archive project',
       error: error.message,
     })
-    send500(res, message)
+    return send500(res, error.message)
   }
-}
-
-export const getUserProjectByIdController = async (req, res) => {
-  const userId = req.session?.user?.id
-  const id = req.params?.id
 
   try {
-    const project = await getUserProjectById(id, userId)
+    // TODO : US #130 appel ansible
+    try {
+      await projectUnlocked(projectId)
 
-    req.log.info({
-      ...getLogInfos({ projectId: project.id }),
-      description: 'Project successfully retrived',
-    })
-    send200(res, project)
+      req.log.info({
+        ...getLogInfos({ projectId }),
+        description: 'Project archived and unlocked',
+      })
+    } catch (error) {
+      req.log.error({
+        ...getLogInfos(),
+        description: 'Cannot unlock project',
+        error: error.message,
+      })
+      return send500(res, error.message)
+    }
   } catch (error) {
-    const message = 'Cannot retrieve project'
     req.log.error({
-      ...getLogInfos({ projectId: id }),
-      description: message,
-      error: error.message,
+      ...getLogInfos(),
+      description: 'Provisioning project with ansible failed',
+      error,
     })
-    send500(res, message)
+    try {
+      await projectFailed(projectId)
+      await projectUnlocked(projectId)
+
+      req.log.info({
+        ...getLogInfos({ projectId }),
+        description: 'Project status successfully updated in database',
+      })
+    } catch (error) {
+      req.log.error({
+        ...getLogInfos(),
+        description: 'Cannot update project status',
+        error: error.message,
+      })
+      return send500(res, error.message)
+    }
+    send500(res, error)
   }
 }
