--- apps/server/src/controllers/project2.js
@@ -1,88 +1,0 @@
-import {
-  projectInitializing,
-  projectCreated,
-  projectFailed,
-  // projectArchiving,
-} from '../models/project-queries2.js'
-import { getLogInfos } from '../utils/logger.js'
-import { send201, send500 } from '../utils/response.js'
-import { ansibleHost, ansiblePort } from '../utils/env.js'
-import { getUserById } from '../models/users-queries.js'
-
-export const createProjectController = async (req, res) => {
-  const data = req.body
-  data.ownerId = req.session.user.id
-
-  let project
-
-  try {
-    project = await projectInitializing(data)
-    req.log.info({
-      ...getLogInfos({
-        projectId: project.id,
-      }),
-      description: 'Project successfully created in database',
-    })
-    send201(res, project)
-  } catch (error) {
-    req.log.error({
-      ...getLogInfos(),
-      description: 'Cannot create project',
-      error: error.message,
-    })
-    return send500(res, error.message)
-  }
-
-  try {
-    const owner = getUserById(project.ownerId)
-    const ansibleData = {
-      orgName: project.organization,
-      ownerEmail: owner.email,
-      projectName: project.name,
-    }
-    await fetch(`http://${ansibleHost}:${ansiblePort}/api/v1/projects`, {
-      method: 'POST',
-      body: JSON.stringify(ansibleData),
-      headers: {
-        'Content-Type': 'application/json',
-        authorization: req.headers.authorization,
-        'request-id': req.id,
-      },
-    })
-
-    try {
-      project = await projectCreated(project.id)
-
-      req.log.info({
-        ...getLogInfos({ projectId: project.id }),
-        description: 'Project status successfully updated in database',
-      })
-    } catch (error) {
-      req.log.error({
-        ...getLogInfos(),
-        description: 'Cannot update project status to created',
-        error: error.message,
-      })
-    }
-  } catch (error) {
-    req.log.error({
-      ...getLogInfos(),
-      description: 'Provisioning project with ansible failed',
-      error,
-    })
-    try {
-      project = await projectFailed(project)
-
-      req.log.info({
-        ...getLogInfos({ projectId: project.id }),
-        description: 'Project status successfully updated in database',
-      })
-    } catch (error) {
-      req.log.error({
-        ...getLogInfos(),
-        description: 'Cannot update project status to failed',
-        error: error.message,
-      })
-    }
-  }
-}
