+++ apps/server/src/models/project-queries.js
@@ -1,151 +1,104 @@
 import { Op } from 'sequelize'
 import { sequelize } from '../connect.js'
 import { getProjectModel } from './models.js'
-import { projectSchema } from 'shared/src/schemas/project.js'
-import { achievedStatus } from 'shared/src/utils/iterables.js'
+import { getUniq } from '../utils/queries-tools.js'
 
-export const createProject = async (project) => {
-  await projectSchema.validateAsync(project)
-  const alreadyExist = await checkUniqueProject(project.orgName, project.projectName)
-  if (alreadyExist) {
-    throw new Error(`Project '${project.orgName}/${project.projectName}' already exists in database`)
-  }
-
-  const res = await getProjectModel().create({ data: project })
-  return res?.data
-}
-
-export const updateProjectStatus = async (project, status) => {
-  project.locked = !achievedStatus.includes(status)
-  await projectSchema.validateAsync(project)
-
-  const res = await getProjectModel().update({
-    data: project,
-  }, {
-    where: {
-      data: {
-        id: {
-          [Op.eq]: project.id,
-        },
-      },
-    },
-  })
-  return res
-}
-
-export const addRepo = async (project, repo) => {
-  if (project.repos?.find(existingRepo => existingRepo.internalRepoName === repo.internalRepoName)) {
-    throw new Error(`Git repo '${repo.internalRepoName}' already exists in project`)
-  }
-
-  project.repos = project.repos?.length ? [...project.repos, repo] : [repo]
-  await projectSchema.validateAsync(project)
-
-  const res = await getProjectModel().update({
-    data: project,
-  }, {
-    where: {
-      data: {
-        id: {
-          [Op.eq]: project.id,
-        },
-      },
-    },
-  })
-  return res
-}
-
-export const addUser = async (project, user) => {
-  if (project.users?.find(existingUser => existingUser.email === user.email) ||
-  project.owner.email === user.email) {
-    throw new Error(`User with email '${user.email}' already member of project`)
-  }
-
-  project.users = project.users?.length ? [...project.users, user] : [user]
-  await projectSchema.validateAsync(project)
-
-  const res = await getProjectModel().update({
-    data: project,
-  }, {
-    where: {
-      data: {
-        id: {
-          [Op.eq]: project.id,
-        },
-      },
-    },
-  })
-  return res
-}
-
-export const removeUser = async (project, userEmail) => {
-  project.users = project.users.filter(user => user.email !== userEmail)
-  await projectSchema.validateAsync(project)
-
-  const res = await getProjectModel().update({
-    data: project,
-  }, {
-    where: {
-      data: {
-        id: {
-          [Op.eq]: project.id,
-        },
-      },
-    },
-  })
-  return res
-}
-
-export const getProjectById = async (id) => {
-  const res = await getProjectModel().findByPk(id)
-  return res
-}
-
-export const getUserProjectById = async (projectId, userId) => {
-  const res = await sequelize.query(
-    'SELECT data FROM "Projects" WHERE (("data"#>>\'{owner,id}\') = $userId OR data->\'users\' @> \'[{"id": "$userId"}]\') AND ("data"#>>\'{id}\') = $projectId LIMIT 1;',
-    {
-      type: sequelize.QueryTypes?.SELECT,
-      bind: { projectId, userId },
-      model: getProjectModel(),
-      plain: true,
-    }).catch(e => { throw e })
-  return res?.data
-}
-
+// SELECT
 export const getUserProjects = async (userId) => {
-  const res = await sequelize.query(
-    'SELECT data FROM "Projects" WHERE (("data"#>>\'{owner,id}\') = $userId OR data->\'users\' @> \'[{"id": "$userId"}]\');',
-    {
-      type: sequelize.QueryTypes?.SELECT,
-      bind: { userId },
-      model: getProjectModel(),
-    }).catch(e => { throw e })
-  return res.map(project => project.data)
-}
-
-export const checkUniqueProject = async (orgName, projectName) => {
-  const res = await getProjectModel().findOne({
+  const res = await getProjectModel().findAll({
+    raw: true,
     where: {
-      [Op.and]: [
-        {
-          data: {
-            projectName: {
-              [Op.iLike]: projectName,
-            },
-          },
-        },
-        {
-          data: {
-            orgName: {
-              [Op.eq]: orgName,
-            },
-          },
-        },
+      [Op.or]: [
+        { ownerId: userId },
+        { usersId: { [Op.contains]: [userId] } },
       ],
     },
   })
   return res
 }
 
-export const _deleteAllProjects = () => getProjectModel().destroy({ truncate: true })
+export const getProjectById = async (id) => {
+  return await getProjectModel().findByPk(id)
+}
+
+export const getProject = async ({ name, organization }) => {
+  const res = await getProjectModel().findAll({
+    raw: true,
+    where: {
+      name,
+      organization,
+    },
+  })
+  return getUniq(res)
+}
+
+// CREATE
+export const projectInitializing = async ({ name, organization, ownerId }) => {
+  const project = await getProject({ name, organization })
+  if (project) throw new Error('Un projet avec le nom et dans l\'organisation demandés existe déjà')
+  const res = await getProjectModel().create({ name, organization, usersId: [ownerId], status: 'initializing', locked: true, ownerId })
+  return res
+}
+
+// UPDATE
+
+export const projectLocked = async (id) => {
+  return await getProjectModel().update({ locked: true }, { where: { id } })
+}
+
+export const projectUnlocked = async (id) => {
+  return await getProjectModel().update({ locked: false }, { where: { id } })
+}
+
+export const projectCreated = async (id) => {
+  return await getProjectModel().update({ locked: false, status: 'created' }, { where: { id } })
+}
+
+export const projectFailed = async (id) => {
+  return await getProjectModel().update({ locked: false, status: 'failed' }, { where: { id } })
+}
+
+export const projectAddUser = async ({ projectId, userId }) => {
+  const project = await getProjectById(projectId)
+  const users = project.usersId ?? [project.ownerId]
+  if (users.includes(userId)) throw new Error('Cet utilisateur est déjà membre du projet')
+  return await getProjectModel().update({
+    usersId: sequelize.fn('array_append', sequelize.col('usersId'), userId),
+  }, {
+    where: { id: projectId },
+  })
+}
+
+// TODO : controller : suppression permissions associées
+export const projectRemoveUser = async ({ projectId, userId }) => {
+  const project = await getProjectById(projectId)
+  const users = project.usersId ?? [project.ownerId]
+  if (!users.includes(userId)) throw new Error('Cet utilisateur n\'est pas membre du projet')
+  return await getProjectModel().update({
+    usersId: sequelize.fn('array_remove', sequelize.col('usersId'), userId),
+  }, {
+    where: { id: projectId },
+  })
+}
+
+export const projectArchiving = async (id) => {
+  const project = await getProjectById(id)
+  if (!project) throw new Error('Projet non trouvé')
+  return await getProjectModel().update({
+    status: 'archived',
+    locked: true,
+  }, {
+    where: { id },
+  })
+}
+
+// DROP
+export const dropProjectsTable = async () => {
+  await sequelize.drop({
+    tableName: getProjectModel().tableName,
+    force: true,
+    cascade: true,
+  })
+}
+
+// DELETE / TRUNCATE
